requirements.txt
agentcoreclient==0.1.8
setproctitle==1.2.2
dnspython

